LayoutTests/imported/w3c/web-platform-tests/css/css-contain/content-visibility/content-visibility-030-expected.txt


FAIL IntersectionObserver interactions assert_equals: runStep2 expected 1 but got 0
